# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ATM Back Office API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portals:
    get:
      operationId: listPortals
      tags: [portals]
      summary: List portal credentials
      responses:
        "200":
          description: Portal credentials (passwords omitted)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Portal"
    post:
      operationId: createPortal
      tags: [portals]
      summary: Add a new portal credential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortalBody"
      responses:
        "201":
          description: Created portal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portal"
  /portals/{id}:
    put:
      operationId: updatePortal
      tags: [portals]
      summary: Update portal credential
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePortalBody"
      responses:
        "200":
          description: Updated portal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portal"
    delete:
      operationId: deletePortal
      tags: [portals]
      summary: Delete portal credential
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /portals/{id}/sync:
    post:
      operationId: syncPortal
      tags: [portals]
      summary: Trigger sync for a specific portal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List planned routes
      responses:
        "200":
          description: Routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
    post:
      operationId: planRoute
      tags: [routes]
      summary: Plan an optimized fill route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanRouteBody"
      responses:
        "201":
          description: Planned route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    get:
      operationId: getRoute
      tags: [routes]
      summary: Get route detail with stops
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Route detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDetail"
    put:
      operationId: updateRoute
      tags: [routes]
      summary: Update route status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRouteBody"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"

  # ─── Alerts ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Portal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          enum: [columbus_data, switch_commerce, atm_transact]
        displayName:
          type: string
        url:
          type: string
        username:
          type: string
        isActive:
          type: boolean
        lastSynced:
          type: string
          format: date-time
        lastSyncStatus:
          type: string
          enum: [success, failed, pending, never]
        atmCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - displayName
        - url
        - username
        - isActive
        - lastSyncStatus
        - createdAt
    CreatePortalBody:
      type: object
      properties:
        name:
          type: string
          enum: [columbus_data, switch_commerce, atm_transact]
        username:
          type: string
        password:
          type: string
      required: [name, username, password]
    UpdatePortalBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        isActive:
          type: boolean
    SyncResult:
      type: object
      properties:
        portalId:
          type: integer
        portalName:
          type: string
        success:
          type: boolean
        message:
          type: string
        atmsUpdated:
          type: integer
        alertsCreated:
          type: integer
        syncedAt:
          type: string
          format: date-time
      required:
        - portalId
        - portalName
        - success
        - message
        - atmsUpdated
        - alertsCreated
        - syncedAt
...
    Route:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        scheduledDate:
          type: string
          format: date
        status:
          type: string
          enum: [planned, in_progress, completed, cancelled]
        stopCount:
          type: integer
        totalCashNeeded:
          type: number
        estimatedDistanceMiles:
          type: number
        estimatedDurationMinutes:
          type: integer
        startAddress:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - scheduledDate
        - status
        - stopCount
        - totalCashNeeded
        - createdAt
    RouteDetail:
      allOf:
        - $ref: "#/components/schemas/Route"
        - type: object
          properties:
            stops:
              type: array
              items:
                $ref: "#/components/schemas/RouteStop"
    RouteStop:
      type: object
      properties:
        id:
          type: integer
        routeId:
          type: integer
        stopOrder:
          type: integer
        atmId:
          type: integer
        atmName:
          type: string
        locationName:
          type: string
        address:
          type: string
        cashToLoad:
          type: number
        currentBalance:
          type: number
        status:
          type: string
          enum: [pending, completed, skipped]
      required:
        - id
        - routeId
        - stopOrder
        - atmId
        - atmName
        - locationName
        - address
        - cashToLoad
        - currentBalance
        - status
    PlanRouteBody:
      type: object
      properties:
        name:
          type: string
        scheduledDate:
          type: string
          format: date
        atmIds:
          type: array
          items:
            type: integer
        daysToFill:
          type: integer
        startAddress:
          type: string
      required: [name, scheduledDate, atmIds, daysToFill]
    UpdateRouteBody:
      type: object
      properties:
        status:
          type: string
          enum: [planned, in_progress, completed, cancelled]

    # ── Alerts ──
...